DataPool size from int to long

+++ RuriLib.Parallelization/ParallelizerFactory.cs
@@ -9,7 +9,7 @@ namespace RuriLib.Parallelization
     {
         public static Parallelizer<TInput, TOutput> Create(ParallelizerType type,
             IEnumerable<TInput> workItems, Func<TInput, CancellationToken, Task<TOutput>> workFunction,
-                int degreeOfParallelism, int totalAmount, int skip = 0)
+                int degreeOfParallelism, long totalAmount, int skip = 0)
         {
             var pType = type switch
             {
